+++ src/main/java/com/furnitureapp/repository/staff/EmployeeRepository.java
@@ -5,7 +5,7 @@ import com.furnitureapp.repository.IRepository;
 
 import java.util.Set;
 
-public interface EmployeeRepository extends IRepository<Employee, Long> {
+public interface EmployeeRepository extends IRepository<Employee, Integer> {
 
     Set<Employee> list();
 }
